# docs/partner-api.openapi.yaml
openapi: 3.1.0
info:
  title: WeChat Channels Demo Partner API
  version: 1.0.0
  description: |
    Server-to-server API for WeChat Channels account containers, QR login,
    hosting state, reply settings, per-account business WeChat QR metadata,
    comments, and direct messages.

    The Bearer credential must be held by the caller's backend. Browser-direct
    access and CORS are not supported. Responses never expose platform
    credentials, raw platform identifiers, concrete model names, upstream
    endpoints, or provider request IDs. Business WeChat QR image bytes are
    accepted only in the settings PUT request and are never returned by Partner
    responses or account projections.
servers:
  - url: https://dev.yytt.com.cn/partner/v1
    description: DEV
security:
  - bearerAuth: []
...
paths:
  /capabilities:
    get:
      tags:
        - Capabilities
      operationId: getPartnerCapabilities
      summary: Get configured reply-provider capabilities
      responses:
        '200':
          description: Stable provider IDs and job-selection capability
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Capabilities'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/PartnerApiUnavailable'
  /accounts:
    post:
      tags:
        - Accounts
      operationId: createPartnerAccount
      summary: Create a disabled account container
      description: |
        Creates a retained account container with automatic replies disabled.
        The request has no body. Repeating the request creates another account.
      responses:
        '201':
          description: Account container created
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountProjection'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          description: Partner API is unavailable or the retained-account limit was reached
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                apiUnavailable:
                  value:
                    error: partner_api_unavailable
                accountLimit:
                  value:
                    error: active_session_limit_reached
    get:
      tags:
        - Accounts
      operationId: listPartnerAccounts
      summary: List all retained accounts
      responses:
        '200':
          description: Retained accounts at every non-deleted login stage
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/PartnerApiUnavailable'
  /accounts/{accountId}:
    parameters:
      - $ref: '#/components/parameters/AccountId'
    get:
      tags:
        - Accounts
      operationId: getPartnerAccount
      summary: Get one account projection
      responses:
        '200':
          description: Current account projection
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountProjection'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '503':
          $ref: '#/components/responses/PartnerApiUnavailable'
    delete:
      tags:
        - Accounts
      operationId: deletePartnerAccount
      summary: Delete an account and its dependent data
      description: |
        Deletes the persisted login credential, comments, direct messages, and
        reply state. The operation is not recoverable. If a platform send is in
        flight, the service preserves the account and returns a conflict.
      responses:
        '204':
          description: Account deleted
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '409':
          description: A platform send is in flight
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: platform_send_in_flight
        '503':
          $ref: '#/components/responses/PartnerApiUnavailable'
  /accounts/{accountId}/login/qr:
    parameters:
      - $ref: '#/components/parameters/AccountId'
    post:
      tags:
        - Login
      operationId: createPartnerLoginQr
      summary: Create or refresh a QR login attempt
      description: |
        A refresh is allowed only for an account that has no valid or
        initializing hosted context. A state conflict preserves all existing
        credentials and content.
      responses:
        '200':
          description: Full account projection with a current QR data URL
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountProjection'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '409':
          description: Login initialization is in progress or the account is already hosted
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                inProgress:
                  value:
                    error: login_in_progress
                alreadyHosted:
                  value:
                    error: account_already_hosted
        '502':
          description: Platform QR creation failed before a login attempt was stored
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: partner_login_qr_unavailable
        '503':
          $ref: '#/components/responses/PartnerApiUnavailable'
  /accounts/{accountId}/login/status:
    parameters:
      - $ref: '#/components/parameters/AccountId'
    get:
      tags:
        - Login
      operationId: getPartnerLoginStatus
      summary: Poll QR scan and login completion state
      description: Poll every two seconds until succeeded or a terminal error state.
      responses:
        '200':
          description: Current login status
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '503':
          $ref: '#/components/responses/PartnerApiUnavailable'
  /accounts/{accountId}/hosting:
    parameters:
      - $ref: '#/components/parameters/AccountId'
    get:
      tags:
        - Hosting
      operationId: getPartnerHosting
      summary: Get hosting and platform-expiry state
      responses:
        '200':
          description: Current hosting projection
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HostingStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '503':
          $ref: '#/components/responses/PartnerApiUnavailable'
    put:
      tags:
        - Hosting
      operationId: updatePartnerHosting
      summary: Enable or pause automatic replies
      description: Source synchronization continues when automatic replies are paused.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HostingUpdate'
      responses:
        '200':
          description: Updated full account projection
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountProjection'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '503':
          $ref: '#/components/responses/PartnerApiUnavailable'
  /accounts/{accountId}/reply-settings:
    parameters:
      - $ref: '#/components/parameters/AccountId'
    put:
      tags:
        - Hosting
      operationId: updatePartnerReplySettings
      summary: Select the reply provider and optional known job number
      description: |
        `funnel` requires a non-empty, caller-supplied `jobNumber`. This API has
        no job catalogue and does not prevalidate whether the job exists.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplySettingsUpdate'
            examples:
              chat:
                value:
                  provider: chat-llm
              funnel:
                value:
                  provider: funnel
                  jobNumber: <known-job-number>
      responses:
        '200':
          description: Updated full account projection
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountProjection'
        '400':
          description: Invalid body or missing Funnel job number
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidRequest:
                  value:
                    error: invalid_request
                jobRequired:
                  value:
                    error: funnel_job_number_required
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '503':
          description: Partner API or selected provider is unavailable
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                apiUnavailable:
                  value:
                    error: partner_api_unavailable
                funnelUnavailable:
                  value:
                    error: funnel_provider_unavailable
  /accounts/{accountId}/wechat-qr:
    parameters:
      - $ref: '#/components/parameters/AccountId'
    get:
      tags:
        - WeChat QR
      operationId: getPartnerWechatQr
      summary: Get business WeChat QR configuration metadata
      description: The response never includes the configured image data URL or Base64 bytes.
      responses:
        '200':
          description: Current QR configuration metadata
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WechatQrSettingsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '503':
          $ref: '#/components/responses/PartnerApiUnavailable'
    put:
      tags:
        - WeChat QR
      operationId: putPartnerWechatQr
      summary: Configure or replace an account business WeChat QR
      description: |
        Accepts one non-empty PNG or JPEG data URL whose decoded bytes are at
        most 512 KiB. The declared MIME must match the image magic bytes. The
        response returns metadata only and never echoes `dataUrl`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WechatQrUpdate'
            example:
              dataUrl: data:image/png;base64,<base64-data>
      responses:
        '200':
          description: Replaced QR configuration metadata without image bytes
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WechatQrSettingsResponse'
        '400':
          description: Invalid image data URL, MIME/signature mismatch, empty image, or decoded image over 512 KiB
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidImage:
                  value:
                    error: account_wechat_qr_invalid
                tooLarge:
                  value:
                    error: account_wechat_qr_too_large
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '503':
          $ref: '#/components/responses/PartnerApiUnavailable'
    delete:
      tags:
        - WeChat QR
      operationId: deletePartnerWechatQr
      summary: Remove an account business WeChat QR
      description: Removes only this account's QR asset and returns unconfigured metadata.
      responses:
        '200':
          description: QR removed; response contains metadata only
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControlNoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WechatQrSettingsResponse'
              example:
                accountId: acc_aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa
                wechatQr:
                  configured: false
                  mimeType: null
                  byteLength: null
                  updatedAt: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/AccountNotFound'
        '503':
          $ref: '#/components/responses/PartnerApiUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
      description: Dedicated Partner API Key held only by the caller's backend.
  headers:
    CacheControlNoStore:
      description: Partner responses must not be cached.
      schema:
        type: string
        const: no-store
...
  responses:
    Unauthorized:
      description: Bearer credential is missing or does not match
      headers:
        Cache-Control:
          $ref: '#/components/headers/CacheControlNoStore'
        WWW-Authenticate:
          description: Authentication scheme required by this API.
          schema:
            type: string
            const: Bearer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: partner_api_unauthorized
    PartnerApiUnavailable:
      description: Partner API credential is not configured on the service
      headers:
        Cache-Control:
          $ref: '#/components/headers/CacheControlNoStore'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: partner_api_unavailable
...
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: string
          minLength: 1
          description: Stable machine-readable error code.
          examples:
            - invalid_request
    Capabilities:
      type: object
      additionalProperties: false
      required:
        - apiVersion
        - providers
        - jobSelection
      properties:
        apiVersion:
          type: string
          const: v1
        providers:
          type: array
          items:
            $ref: '#/components/schemas/ProviderCapability'
        jobSelection:
          $ref: '#/components/schemas/JobSelectionCapability'
      example:
        apiVersion: v1
        providers:
          - id: chat-llm
            displayName: CHAT回复
            configured: true
            requiresJobNumber: false
          - id: funnel
            displayName: 招聘接口
            configured: true
            requiresJobNumber: true
        jobSelection:
          mode: known_job_number
          catalogueAvailable: false
...
    AccountList:
      type: object
      additionalProperties: false
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AccountProjection'
...